jd_parser/requirements_jd_parser.txt
langchain-huggingface
langchain-core
pydantic
python-dotenv
huggingface-hub
